requirements.txt
torch
ultralytics
streamlit
opencv-python-headless
pillow
numpy
